requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
pytesseract
pdf2image
numpy
pydantic
requests
pytest
httpx
google-generativeai
python-dotenv
dotenv